# docs/swagger.yaml
basePath: /v1
definitions:
  entity.AccessToken:
    properties:
      access_token:
        type: string
    type: object
...
  entity.ClientType:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  entity.LoginRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 16
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  entity.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 30
        minLength: 2
        type: string
      last_name:
        type: string
      password:
        maxLength: 16
        minLength: 6
        type: string
      phone_number:
        type: string
    required:
    - email
    - first_name
    - password
    type: object
  entity.Role:
    properties:
      client_type_data:
        $ref: '#/definitions/entity.ClientType'
      client_type_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  entity.UpdatePasswordRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  entity.User:
    properties:
      access_token:
        type: string
      business_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      refresh_token:
        type: string
      role_data:
        $ref: '#/definitions/entity.Role'
      role_id:
        type: string
      updated_at:
        type: string
    type: object
  entity.VerifyEmail:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
  entity.VerifyRequest:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        type: string
    required:
    - code
    - email
    - new_password
    type: object
  status_http.Response:
    properties:
      custom_message: {}
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: Go Clean Template API
  version: "1.0"
paths:
  /auth/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete account with id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete account
      tags:
      - AUTH
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Forgot password
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Forgot password
      tags:
      - AUTH
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login by email and password
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/entity.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login
      tags:
      - AUTH
  /auth/refresh-token/:token:
    post:
      consumes:
      - application/json
      description: Refresh access-token
      parameters:
      - description: Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.AccessToken'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Creates new valid access-token
      tags:
      - AUTH
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register
      parameters:
      - description: RegisterRequest
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/entity.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status_http.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Register
      tags:
      - AUTH
  /auth/update-password:
    post:
      consumes:
      - application/json
      description: Update password
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.UpdatePasswordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update password
      tags:
      - AUTH
  /auth/verify:
    post:
      consumes:
      - application/json
      description: Verify user
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.VerifyEmail'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/status_http.Response'
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Verify user
      tags:
      - AUTH
  /auth/verify-forgot-password:
    post:
      consumes:
      - application/json
      description: Verify forgot password
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.VerifyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/status_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Verify forgot password
      tags:
      - AUTH
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"